pyproject.toml
[tool.poetry]
name = "djcraft"
version = "0.1.0"
description = "Boilerplate generator for django projects."
authors = ["munjed-ab <abazamunged@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Jinja2 = "^3.0.3"
click = "^8.0.4"
rich = "^14.0.0"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
django-boilerplate = "boilerplate_generator.cli:main"
